pyproject.toml
[tool.poetry]
name = "minmodkg"
version = "1.2.0"
description = "code repository for TA2 Knowledge Graph (MinMod) construction & deployment"
authors = ["Binh Vu <bvu687@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "minmodkg" }]

[tool.poetry.dependencies]
python = "^3.11"
statickg = "^1.9.2"
fastapi = "^0.111.0"
networkx = "^2.8.8"
httpx = "^0.27.0"
htbuilder = "^0.6.2"
pandas = "^2.2.2"
shapely = "^2.0.5"
pyproj = "^3.6.1"
pyjwt = "^2.9.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }

[tool.poetry.dev-dependencies]
erdantic = "^1.0.3"
pydantic = "^2.7.4"
pytest = "^8.3.2"
jsonschema = "^4.23.0"
pytest-dotenv = "^0.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
